Add files via upload
It finds speccific comments, hashtags and emoticons for account and use it when running. It locate speccific fields and insert or click on speciffic data. The idea is to open imgage click follow account, like photo/video and write speccific comment, then add emoticon and go to the next post. The program repeat this process in loop.